dev-requirements.txt
pytest==2.5.1
Sphinx==1.1.3
coverage
python-coveralls
python-dateutil==2.2
